# Normalize comparisons and improve string_t operations