--- a/src/common/types/string_heap.cpp
+++ b/src/common/types/string_heap.cpp
@@ -50,7 +50,7 @@ string_t StringHeap::AddBlob(const string_t &data) {
 }
 
 string_t StringHeap::EmptyString(idx_t len) {
-	D_ASSERT(len >= string_t::INLINE_LENGTH);
+	D_ASSERT(len > string_t::INLINE_LENGTH);
 	auto insert_pos = (const char *)allocator.Allocate(len);
 	return string_t(insert_pos, len);
 }