--- a/src/include/duckdb/common/types/string_type.hpp
+++ b/src/include/duckdb/common/types/string_type.hpp
@@ -173,8 +173,7 @@ public:
 				return A > B;
 #endif
 			auto memcmp_res = memcmp(left.GetDataUnsafe(), right.GetDataUnsafe(), min_length);
-			auto final_res = (memcmp_res == 0) ? (left_length > right_length) : (memcmp_res > 0);
-			return final_res;
+			return memcmp_res > 0 || (memcmp_res == 0 && left_length > right_length);
 		}
 	};