… src/include/duckdb/common/types/interval.hpp
… test/sql/types/interval/test_interval_comparison.test
… test/sql/window/test_window_range.test