This should only have small positive effects on 2 areas:

- 2-3% binary size reduction (varies on included extensions) on duckdb binary
This is thanks to folding templates into equivalent operators, thanks to the fact that:
```
a >  b --> a > b
a <  b --> b > a
a >= b --> a > b, negated
a <= b --> b > a, negated
```
Changing order of operations is sort of possible in all cases, negating the comparison might need some rework of the calling code like
```
count(a>=b) --> size - count(a>b)
```
that has been partially done, but probably with some effort can be done in even more places.
Secondary advantage is in code reduction.
Test are all passing but one tied to interval >= logic, where given that `==` is different than `nor > nor <', where we move from `>= is > or ==` to a behaviour compliant with postgres. (test is [test/sql/types/interval/test_interval_comparison.test](https://github.com/duckdb/duckdb/compare/feature...carlopi:normalizeComparisons1?expand=1#diff-1749fa6a9596d67becfebc100121b5425d04410f512633f65d4a027408624e16)).

I used [bloaty](https://github.com/google/bloaty), a binary size profiler, to detect relevant highly duplicated templates. Next up are functions like those (and their permutations with other types) but that requires swapping also output-parameters.
```
   0.1%  21.6Ki   0.1%  21.6Ki    unsigned long long duckdb::DistinctSelectGeneric<duckdb::string_t, duckdb::string_t, duckdb::DistinctGreaterThanNullsFirst>(duckdb::Vector&, duckdb::Vector&, duckdb::SelectionVector const*, unsigned long long, duckdb::SelectionVector*, duckdb::SelectionVector*)
   0.1%  21.6Ki   0.1%  21.6Ki    unsigned long long duckdb::DistinctSelectGeneric<duckdb::string_t, duckdb::string_t, duckdb::DistinctGreaterThan>(duckdb::Vector&, duckdb::Vector&, duckdb::SelectionVector const*, unsigned long long, duckdb::SelectionVector*, duckdb::SelectionVector*)
   0.1%  21.6Ki   0.1%  21.6Ki    unsigned long long duckdb::DistinctSelectGeneric<duckdb::string_t, duckdb::string_t, duckdb::DistinctGreaterThanEquals>(duckdb::Vector&, duckdb::Vector&, duckdb::SelectionVector const*, unsigned long long, duckdb::SelectionVector*, duckdb::SelectionVector*)
   0.1%  21.6Ki   0.1%  21.6Ki    unsigned long long duckdb::DistinctSelectGeneric<duckdb::string_t, duckdb::string_t, duckdb::DistinctLessThanNullsFirst>(duckdb::Vector&, duckdb::Vector&, duckdb::SelectionVector const*, unsigned long long, duckdb::SelectionVector*, duckdb::SelectionVector*)
   ```

- perf improvements in string_t related equality, greater-than or hashing functions. In string heavy workloads this should be worth up to 10% improvement (handwavy).
Here we using the earlier simplification to deal with simpler code, and we use the fact that string_t have always 12 bytes of valid access to speed up greater than operator and hashing, and using more heavily the fact that binary equal `string_t`s have to be equal (and the converse it's also true in limited cases, like prefix being different).

Here some hashing-related tests have to be modified for allowing faster computation (expecially on inlined strings where we avoid function calls / and pay only 1 branch to detect inlining).

I believe (should add a test) that in this version hash(string_t) should be the same of the equivalent hash(char*).

- added passing by `const &` in the comparison_operator interface
This should be negligible, but I believe it's cleaner and might, in some rare circumstances, enable better code generation.