@Mytherin 2023-02-20
Also, it looks like TPC-H Q1 (which uses string hashing and comparisons very intensely) got significantly slower:

```
benchmark/tpch/sf1/q01.benchmark
Old timing: 0.180383
New timing: 0.227067
```

Perhaps have a look there and see if this new comparison/hash actually improves things?