@carlopi 2023-02-28
> Also, it looks like TPC-H Q1 (which uses string hashing and comparisons very intensely) got significantly slower:
> ...
> ```
> benchmark/tpch/sf1/q01.benchmark
> Old timing: 0.180383
> New timing: 0.227067
> ```
> 
> Perhaps have a look there and see if this new comparison/hash actually improves things?

Actually I added back an optimisation that I though was not needed, and now numbers look like this:

CI (there is lots of noise, but seems to be faster)
```
benchmark/tpch/sf1/q01.benchmark
Old timing: 0.16775
New timing: 0.164793

benchmark/tpch/sf1/q19.benchmark
Old timing: 0.145454
New timing: 0.129259
```

MacOS M1 (in all those cases this PR is faster that base branch by at least a few sigma, plus a bunch of tests with unchanged results)
```
micro/cast/cast_date_string.benchmark         a = 0.025s      b = 0.036s      ( 0.010917s     , 30.52%        , 58.020 sigma)
micro/cast/cast_int_string.benchmark          a = 0.024s      b = 0.032s      ( 0.007647s     , 24.23%        , 21.385 sigma)
micro/cast/cast_double_string.benchmark       a = 0.106s      b = 0.118s      ( 0.012924s     , 10.91%        , 33.491 sigma)
micro/cast/cast_timestamp_string.benchmark    a = 0.042s      b = 0.058s      ( 0.016011s     , 27.61%        , 77.429 sigma)
tpch/sf1/q01.benchmark                        a = 0.039s      b = 0.039s      ( 0.000774s     , 1.96%         , 2.494 sigma)
tpch/sf1/q19.benchmark                        a = 0.030s      b = 0.033s      ( 0.003012s     , 9.22%         , 4.262 sigma)
tpcds/sf1/q02.benchmark                       a = 0.028s      b = 0.030s      ( 0.001700s     , 5.66%         , 3.575 sigma)
tpcds/sf1/q04.benchmark                       a = 0.135s      b = 0.146s      ( 0.010830s     , 7.42%         , 6.499 sigma)
tpcds/sf1/q11.benchmark                       a = 0.110s      b = 0.121s      ( 0.011397s     , 9.41%         , 7.115 sigma)
tpcds/sf1/q22.benchmark                       a = 0.132s      b = 0.137s      ( 0.004544s     , 3.32%         , 4.712 sigma)
tpcds/sf1/q36.benchmark                       a = 0.035s      b = 0.039s      ( 0.003493s     , 8.97%         , 4.016 sigma)
tpcds/sf1/q59.benchmark                       a = 0.049s      b = 0.052s      ( 0.002452s     , 4.72%         , 3.571 sigma)
tpcds/sf1/q74.benchmark                       a = 0.063s      b = 0.069s      ( 0.006030s     , 8.72%         , 4.055 sigma)
```

I think this is worth putting in, since fast-path for string_type operations, expecially if inlined, seems to be significant enough.

There are 2 unfinished parts:
* integrate A/B testing script with current regression testing script
This I will do, separating A/B script to a separate PR.

* there is any decent way around *reinterpret_cast<uint32_t>(some_pointer) ?
Load<uint32_t> do not currently propagates alignment informations, and the compilers seems not to be able to recreate them in all cases. On platforms where the penalty for unaligned loads is negligible, this is not a problem, and performances are the same between `Load<>` and `*reinterpret_cast<>`, but in architectures where this is a hard failure, the fact that no alignment information is not propagated is a bit bump (say 30% in tight loops). Here we are leveraging the fact that string_type will always be allocated at least aligned at 4 bytes, so it's members are always aligned to 4 bytes.

(There have been also some unconnected failures in CI, it passed but TPCDS on my fork)